pyproject.toml
[project]
name = "hg-chat-with-model"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikolai Polushkin",email = "mumig12@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub (>=1.2.4,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "hg_chat_with_model", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
